+++ src/app/static/src/app/featureSwitches.ts
@@ -1,8 +1,8 @@
 const urlParams = new URLSearchParams(window.location.search);
-const adrPushInputs = urlParams.get('adrPushInputs');
+const modelBugReport = !urlParams.get('modelBugReport');
 const modelCalibratePlot = urlParams.get('modelCalibratePlot');
 
 export const switches = {
-    adrPushInputs,
+    modelBugReport,
     modelCalibratePlot
 };